Refactor to move RateLimiter classes to org.apache.flink.api.common.io
