+++ flink-core/pom.xml
@@ -80,6 +80,12 @@ under the License.
 			<!-- managed version -->
 		</dependency>
 
+		<!-- Ratelimiting dependencies -->
+		<dependency>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-shaded-guava</artifactId>
+		</dependency>
+
 		<!-- ================== test dependencies ================== -->
 
 		<dependency>
