+++ flink-core/src/main/java/org/apache/flink/api/common/io/ratelimiting/GuavaFlinkConnectorRateLimiter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.connectors.kafka.config;
+package org.apache.flink.api.common.io.ratelimiting;
 
 import org.apache.flink.api.common.functions.RuntimeContext;
